+++ SoundEffects/Assets/SoundEffects/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.2.0] - 2024-07-29
 ### Added
 - Added a feature to automatically search for Detail assets.
+### Changed
+- Changed to pass the SoundEffectPlayRequest structure when playing sound effects.
 
 ## [0.1.1] - 2024-05-25
 ### Added